requirements.txt
crispy-bootstrap5==0.7
Django==4.1.2
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-environ==0.9.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.4
